Aula 8 - Adicionando script start no package.json para iniciar a aplicacao
