+++ src/app.js
@@ -1,0 +1,16 @@
+'use strict';
+
+const express = require('express');
+
+const app = express();
+const router = express.Router();
+
+const route = router.get('/', (req, res, next) => {
+    res.status(200).send({
+        title: "Estudando API Nodejs",
+        version: "0.0.1"
+    });
+});
+app.use('/', route);
+
+module.exports = app;
